requirements.txt
aiogram~=3.0.0b8
python-dotenv~=1.0.0